Separated calculation logic from the component.
